Move graphql-related functions into graphql directory
