+++ src/graphql/fetchDataSubjects.ts
