+++ src/graphql/index.ts
@@ -1,1 +1,8 @@
 export * from './fetchActiveSiloDiscoPlugin';
+export * from './fetchApiKeys';
+export * from './fetchDataSubjects';
+export * from './fetchIdentifiers';
+export * from './pullTranscendConfiguration';
+export * from './syncConfigurationToTranscend';
+export * from './syncDataSilos';
+export * from './syncEnrichers';
