+++ src/graphql/syncConfigurationToTranscend.ts
